requirements.txt
beautifulsoup4==4.4.1
click==6.6
decorator==4.0.10
dj-database-url==0.4.1
Django==1.9.7
django-autoslug==1.9.3
django-localflavor==1.3
future==0.15.2
geocoder==1.15.1
gunicorn==19.6.0
Metafone==0.5
Pillow==3.3.0
probableparsing==0.0.1
psycopg2==2.6.2
python-crfsuite==0.8.4
ratelim==0.1.6
requests==2.10.0
six==1.10.0
usaddress==0.5.7
whitenoise==3.2
